Added tests for the REST endpoints exposed in our example. Added a script for running the example app. Split the example into multiple files.
